# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/admin/api/v1/api-keys": {
            "get": {
                "security": [
                    {
                        "AdminJWT": []
                    }
                ],
                "description": "Returns all API keys",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API Keys"
                ],
                "summary": "List API keys",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/mdb.ApiKey"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "AdminJWT": []
                    }
                ],
                "description": "Create a new universal API key. Usable for both gateway flows (epay/ezpay). PID auto-incremented; secret returned once.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API Keys"
                ],
                "summary": "Create API key",
                "parameters": [
                    {
                        "description": "API key payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.CreateApiKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/admin.CreateApiKeyResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/admin/api/v1/api-keys/{id}": {
            "delete": {
                "security": [
                    {
                        "AdminJWT": []
                    }
                ],
                "description": "Soft-delete an API key",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API Keys"
                ],
                "summary": "Delete API key",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "API key ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "AdminJWT": []
                    }
                ],
                "description": "Patch API key name / ip_whitelist / notify_url",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API Keys"
                ],
                "summary": "Update API key",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "API key ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Fields to update",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.UpdateApiKeyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/admin/api/v1/api-keys/{id}/rotate-secret": {
            "post": {
                "security": [
                    {
                        "AdminJWT": []
                    }
                ],
                "description": "Generate a new secret_key. Returned only once.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API Keys"
                ],
                "summary": "Rotate API key secret",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "API key ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/admin.RotateSecretResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/admin/api/v1/api-keys/{id}/secret": {
            "get": {
                "security": [
                    {
                        "AdminJWT": []
                    }
                ],
                "description": "Returns the secret_key for an API key (admin only)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API Keys"
                ],
                "summary": "View API key secret",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "API key ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/admin/api/v1/api-keys/{id}/stats": {
            "get": {
                "security": [
                    {
                        "AdminJWT": []
                    }
                ],
                "description": "Returns usage stats (call_count, last_used_at) for an API key",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API Keys"
                ],
                "summary": "Get API key stats",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "API key ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/admin.ApiKeyStatsResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/admin/api/v1/api-keys/{id}/status": {
            "post": {
                "security": [
                    {
                        "AdminJWT": []
                    }
                ],
                "description": "Toggle enable/disable for an API key (1=enabled, 2=disabled)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin API Keys"
                ],
                "summary": "Change API key status",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "API key ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.ChangeApiKeyStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/admin/api/v1/auth/init-password": {
            "get": {
                "description": "Returns the initial random admin password once, then invalidates it.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Auth"
                ],
                "summary": "Get initial admin password (one-time)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/admin.InitialPasswordResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/admin/api/v1/auth/init-password-hash": {
            "get": {
                "description": "Returns the initial password hash and password-changed flag.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Auth"
                ],
                "summary": "Get initial admin password hash",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/data.InitialAdminPasswordHashInfo"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/admin/api/v1/auth/login": {
            "post": {
                "description": "Verify credentials and return a signed JWT token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Auth"
                ],
                "summary": "Admin login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/admin.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/admin.LoginResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/admin/api/v1/auth/logout": {
            "post": {
                "security": [
                    {
                        "AdminJWT": []
                    }
                ],
                "description": "Logout (no-op, frontend discards token)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Auth"
                ],
                "summary": "Admin logout",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
        "/admin/api/v1/auth/me": {
            "get": {
                "security": [
                    {
                        "AdminJWT": []
                    }
                ],
                "description": "Returns the currently authenticated admin user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin Auth"
                ],
                "summary": "Get current admin profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/admin.MeResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ApiResponse"
                        }
                    }
                }
            }
        },
...
        "/pay/checkout-counter-resp/{trade_id}": {
            "get": {
                "description": "Render the payment checkout counter JSON response for a given trade",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Checkout counter page",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Trade ID",
                        "name": "trade_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.CheckoutCounterResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "admin.CreateApiKeyRequest": {
            "type": "object",
            "properties": {
                "ip_whitelist": {
                    "type": "string",
                    "example": ""
                },
                "name": {
                    "type": "string",
                    "example": "My API Key"
                },
                "notify_url": {
                    "type": "string",
                    "example": "https://example.com/notify"
                }
            }
        },
        "admin.CreateApiKeyResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "My API Key"
                },
                "pid": {
                    "type": "string",
                    "example": "1003"
                },
                "secret_key": {
                    "description": "SecretKey is returned ONCE on creation. After that, fetch via\nGET /api-keys/:id/secret or rotate to generate a new one.",
                    "type": "string",
                    "example": "secret123abc456"
                }
            }
        },
...
        "response.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success"
                },
                "request_id": {
                    "type": "string",
                    "example": "req-123456"
                },
                "status_code": {
                    "type": "integer",
                    "example": 200
                }
            }
        },
...
    }
}